apps/api/requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-genai
groq
pytest
